Add two splutions "Pig-style text" and "Readable time format"
Pigstyle text: I need to move the first letter of each word to the end of it, then add \"ay\" to the end of the word.
Readable time format: task in order to complete this Kata is to write a function which formats a duration, given as a number of seconds, in a human-friendly way. The duration is expressed as a combination of years, days, hours, minutes and seconds.